Et_tech/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.7.1
Django==5.1
django-crispy-forms==2.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
httpie==3.2.3
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
oauthlib==3.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==72.1.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
